Renaming client and server libraries
