+++ client/python/setup.cfg
@@ -4,14 +4,14 @@
 # https://setuptools.pypa.io/en/latest/references/keywords.html
 
 [metadata]
-name = djclient
-description = DJ client library
+name = datajunction
+description = DataJunction Client Library
 author = DataJunction
 license = MIT
 license_files = LICENSE.txt
 long_description = file: README.md
 long_description_content_type = text/x-markdown; charset=UTF-8
-url = https://github.com/DataJunction/dj/
+url = https://github.com/DataJunction/dj/tree/main/client/python
 
 project_urls =
     Source = https://github.com/DataJunction/dj/
@@ -31,7 +31,7 @@ zip_safe = False
 packages = find_namespace:
 include_package_data = True
 package_dir =
-    =djclient
+    =datajunction
 
 # Require a min/specific Python version (comma-separated conditions)
 python_requires = >=3.8
@@ -46,13 +46,13 @@ install_requires =
     pandas>=2
 
 [options.packages.find]
-where = djclient
+where = datajunction
 exclude =
     tests
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
-# `pip install djclient[PDF]` like:
+# `pip install datajunction[PDF]` like:
 # PDF = ReportLab; RXP
 
 # Add here test requirements (semicolon/line-separated)
@@ -65,10 +65,10 @@ testing =
 [options.entry_points]
 # Add here console scripts like:
 # console_scripts =
-#     script_name = djclient.module:function
+#     script_name = datajunction.module:function
 # For example:
 # console_scripts =
-#     fibonacci = djclient.skeleton:run
+#     fibonacci = datajunction.skeleton:run
 # And any other entry points, for example:
 # pyscaffold.cli =
 #     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
@@ -80,7 +80,7 @@ testing =
 # CAUTION: --cov flags may prohibit setting breakpoints while debugging.
 #          Comment those flags to avoid this pytest issue.
 addopts =
-    --cov djclient --cov-report term-missing
+    --cov datajunction --cov-report term-missing
     --verbose
 norecursedirs =
     dist
@@ -115,4 +115,4 @@ exclude =
 # PyScaffold's parameters when the project was created.
 # This will be used when updating. Do not change!
 version = 4.4
-package = djclient
+package = datajunction
