+++ docs/content/0.1.0/docs/getting-started/creating-nodes/sources.md
@@ -48,8 +48,9 @@ curl -X POST http://localhost:8000/nodes/source/ \
 ```
 {{< /tab >}}
 {{< tab "python" >}}
+
 ```py
-from djclient import DJClient, NodeMode
+from datajunction import DJClient, NodeMode
 
 dj = DJClient("http://localhost:8000/")
 source = dj.new_source(
@@ -95,8 +96,9 @@ curl -X POST http://localhost:8000/nodes/source/ \
 ```
 {{< /tab >}}
 {{< tab "python" >}}
+
 ```py
-from djclient import DJClient, NodeMode
+from datajunction import DJClient, NodeMode
 
 dj = DJClient("http://localhost:8000/")
 source = dj.new_source(
