+++ docs/content/0.1.0/docs/getting-started/creating-nodes/transforms.md
@@ -43,8 +43,9 @@ curl -X POST http://localhost:8000/nodes/transform/ \
 ```
 {{< /tab >}}
 {{< tab "python" >}}
+
 ```py
-from djclient import DJClient, NodeMode
+from datajunction import DJClient, NodeMode
 
 dj = DJClient("http://localhost:8000/")
 transform = dj.new_transform(
