+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 version_scheme = "no-guess-dev"
 
 [tool.poetry]
-name = "datajunction"
+name = "datajunction-server"
 version = "0.0.1a5"
 readme = "README.rst"
 repository = "https://github.com/DataJunction/dj"
